+++ packages/client/src/object/object.test.ts
@@ -15,9 +15,10 @@
  */
 
 import type { Osdk } from "@osdk/client.api";
-import { $ontologyRid, Employee } from "@osdk/client.test.ontology";
+import { $Objects, $ontologyRid, Employee } from "@osdk/client.test.ontology";
 import { apiServer, stubData, withoutRid } from "@osdk/shared.test";
 import { afterAll, beforeAll, describe, expect, it } from "vitest";
+import type { UnstableClient } from "../__unstable/UnstableClient.js";
 import type { Client } from "../Client.js";
 import { createClient } from "../createClient.js";
 
@@ -151,24 +152,39 @@ describe("OsdkObject", () => {
         "ri.phonograph2-objects.main.object.88a6fccb-f333-46d6-a07e-7725c5f18b61",
       );
     });
+    it("objects are enumerable in an sdk", async () => {
+      const objects = Object.keys($Objects);
+      expect(objects).toStrictEqual([
+        "Employee",
+        "ObjectWithTimestampPrimaryKey",
+        "Office",
+        "Person",
+        "Task",
+        "Todo",
+        "equipment",
+        "objectTypeWithAllPropertyTypes",
+      ]);
+    });
   });
 });
 
-async function shouldError(client: Client): Promise<Osdk<Employee>> {
+export async function shouldError(client: Client): Promise<Osdk<Employee>> {
   // @ts-expect-error
   return client(Employee).fetchOne(1, {
     $select: ["employeeId"],
   });
 }
 
-async function shouldError2(client: Client): Promise<Employee.OsdkObject> {
+export async function shouldError2(
+  client: Client,
+): Promise<Employee.OsdkObject> {
   // @ts-expect-error
   return client(Employee).fetchOne(1, {
     $select: ["employeeId"],
   });
 }
 
-async function shouldCompile(
+export async function shouldCompile_client_fetchOne_old_select(
   client: Client,
 ): Promise<Osdk<Employee, "employeeId">> {
   return client(Employee).fetchOne(1, {
@@ -176,10 +192,38 @@ async function shouldCompile(
   });
 }
 
-async function shouldCompile2(
+export async function shouldCompile_unstableClient_fetchOne_old_select(
+  client: UnstableClient,
+): Promise<Osdk<Employee, "employeeId">> {
+  return client(Employee).fetchOne(1, {
+    $select: ["employeeId"],
+  });
+}
+
+export async function shouldCompile_client_fetchOne_new_select(
   client: Client,
 ): Promise<Employee.OsdkObject<never, "employeeId">> {
   return client(Employee).fetchOne(1, {
     $select: ["employeeId"],
   });
+}
+
+export async function shouldCompile_unstableClient_fetchOne_new_select(
+  client: UnstableClient,
+): Promise<Osdk<Employee, "employeeId">> {
+  return client(Employee).fetchOne(1, {
+    $select: ["employeeId"],
+  });
+}
+
+export async function shouldCompile_client_fetchOne_old_noArgs(
+  client: Client,
+): Promise<Osdk<Employee>> {
+  return client(Employee).fetchOne(1);
+}
+
+export async function shouldCompile_unstableClient_fetchOne_noArgs(
+  client: UnstableClient,
+): Promise<Osdk<Employee>> {
+  return client(Employee).fetchOne(1);
 }