+++ packages/generator/src/v2.0/UNSTABLE_wireInterfaceTypeV2ToSdkObjectConst.test.ts
@@ -113,6 +113,7 @@ describe(__UNSTABLE_wireInterfaceTypeV2ToSdkObjectConst, () => {
         DefaultToFalse as $DefaultToFalse,
         FetchPageArgs as $FetchPageArgs,
         FetchPageResult as $FetchPageResult,
+        IsAny as $IsAny,
         LinkedType as $LinkedType,
         LinkNames as $LinkNames,
         NullabilityAdherence as $NullabilityAdherence,
@@ -141,7 +142,7 @@ describe(__UNSTABLE_wireInterfaceTypeV2ToSdkObjectConst, () => {
         InterfaceDefinition as $InterfaceDefinition,
       } from "@osdk/api";
 
-      export type OsdkObjectLinks$Bar = never;
+      export type OsdkObjectLinks$Bar = {};
 
       export namespace Bar {
         export type PropertyKeys = "bar";
@@ -174,7 +175,7 @@ describe(__UNSTABLE_wireInterfaceTypeV2ToSdkObjectConst, () => {
               Bar.OsdkObject<
                 | (S extends false ? "$notStrict" : never)
                 | ($DefaultToFalse<R> extends false ? never : "$rid"),
-                L
+                $IsAny<L> extends true ? Bar.PropertyKeys : L
               >
             >
           >;
@@ -192,7 +193,7 @@ describe(__UNSTABLE_wireInterfaceTypeV2ToSdkObjectConst, () => {
                 Bar.OsdkObject<
                   | (S extends false ? "$notStrict" : never)
                   | ($DefaultToFalse<R> extends false ? never : "$rid"),
-                  L
+                  $IsAny<L> extends true ? Bar.PropertyKeys : L
                 >
               >
             >
@@ -207,6 +208,7 @@ describe(__UNSTABLE_wireInterfaceTypeV2ToSdkObjectConst, () => {
           osdkMetadata: typeof $osdkMetadata;
           objectSet: Bar.ObjectSet;
           props: Bar.Props;
+          linksType: OsdkObjectLinks$Bar;
           strictProps: Bar.StrictProps;
           displayName: "Bar";
           implements: [];
@@ -248,6 +250,7 @@ describe(__UNSTABLE_wireInterfaceTypeV2ToSdkObjectConst, () => {
         osdkMetadata: $osdkMetadata,
         objectSet: undefined as any,
         props: undefined as any,
+        linksType: undefined as any,
         strictProps: undefined as any,
         apiName: "Bar",
         displayName: "Bar",
@@ -301,6 +304,7 @@ describe(__UNSTABLE_wireInterfaceTypeV2ToSdkObjectConst, () => {
         DefaultToFalse as $DefaultToFalse,
         FetchPageArgs as $FetchPageArgs,
         FetchPageResult as $FetchPageResult,
+        IsAny as $IsAny,
         LinkedType as $LinkedType,
         LinkNames as $LinkNames,
         NullabilityAdherence as $NullabilityAdherence,
@@ -329,7 +333,7 @@ describe(__UNSTABLE_wireInterfaceTypeV2ToSdkObjectConst, () => {
         InterfaceDefinition as $InterfaceDefinition,
       } from "@osdk/api";
 
-      export type OsdkObjectLinks$Foo = never;
+      export type OsdkObjectLinks$Foo = {};
 
       export namespace Foo {
         export type PropertyKeys = "foo";
@@ -365,7 +369,7 @@ describe(__UNSTABLE_wireInterfaceTypeV2ToSdkObjectConst, () => {
               Foo.OsdkObject<
                 | (S extends false ? "$notStrict" : never)
                 | ($DefaultToFalse<R> extends false ? never : "$rid"),
-                L
+                $IsAny<L> extends true ? Foo.PropertyKeys : L
               >
             >
           >;
@@ -383,7 +387,7 @@ describe(__UNSTABLE_wireInterfaceTypeV2ToSdkObjectConst, () => {
                 Foo.OsdkObject<
                   | (S extends false ? "$notStrict" : never)
                   | ($DefaultToFalse<R> extends false ? never : "$rid"),
-                  L
+                  $IsAny<L> extends true ? Foo.PropertyKeys : L
                 >
               >
             >
@@ -398,6 +402,7 @@ describe(__UNSTABLE_wireInterfaceTypeV2ToSdkObjectConst, () => {
           osdkMetadata: typeof $osdkMetadata;
           objectSet: Foo.ObjectSet;
           props: Foo.Props;
+          linksType: OsdkObjectLinks$Foo;
           strictProps: Foo.StrictProps;
           description: "Foo interface desc";
           displayName: "Foo interface dn";
@@ -441,6 +446,7 @@ describe(__UNSTABLE_wireInterfaceTypeV2ToSdkObjectConst, () => {
         osdkMetadata: $osdkMetadata,
         objectSet: undefined as any,
         props: undefined as any,
+        linksType: undefined as any,
         strictProps: undefined as any,
         apiName: "Foo",
         description: "Foo interface desc",
@@ -503,6 +509,7 @@ describe(__UNSTABLE_wireInterfaceTypeV2ToSdkObjectConst, () => {
         DefaultToFalse as $DefaultToFalse,
         FetchPageArgs as $FetchPageArgs,
         FetchPageResult as $FetchPageResult,
+        IsAny as $IsAny,
         LinkedType as $LinkedType,
         LinkNames as $LinkNames,
         NullabilityAdherence as $NullabilityAdherence,
@@ -531,7 +538,7 @@ describe(__UNSTABLE_wireInterfaceTypeV2ToSdkObjectConst, () => {
         InterfaceDefinition as $InterfaceDefinition,
       } from "@osdk/api";
 
-      export type OsdkObjectLinks$Foo = never;
+      export type OsdkObjectLinks$Foo = {};
 
       export namespace Foo {
         export type PropertyKeys = "foo" | "bar";
@@ -569,7 +576,7 @@ describe(__UNSTABLE_wireInterfaceTypeV2ToSdkObjectConst, () => {
               Foo.OsdkObject<
                 | (S extends false ? "$notStrict" : never)
                 | ($DefaultToFalse<R> extends false ? never : "$rid"),
-                L
+                $IsAny<L> extends true ? Foo.PropertyKeys : L
               >
             >
           >;
@@ -587,7 +594,7 @@ describe(__UNSTABLE_wireInterfaceTypeV2ToSdkObjectConst, () => {
                 Foo.OsdkObject<
                   | (S extends false ? "$notStrict" : never)
                   | ($DefaultToFalse<R> extends false ? never : "$rid"),
-                  L
+                  $IsAny<L> extends true ? Foo.PropertyKeys : L
                 >
               >
             >
@@ -602,6 +609,7 @@ describe(__UNSTABLE_wireInterfaceTypeV2ToSdkObjectConst, () => {
           osdkMetadata: typeof $osdkMetadata;
           objectSet: Foo.ObjectSet;
           props: Foo.Props;
+          linksType: OsdkObjectLinks$Foo;
           strictProps: Foo.StrictProps;
           description: "Foo interface desc";
           displayName: "Foo interface dn";
@@ -650,6 +658,7 @@ describe(__UNSTABLE_wireInterfaceTypeV2ToSdkObjectConst, () => {
         osdkMetadata: $osdkMetadata,
         objectSet: undefined as any,
         props: undefined as any,
+        linksType: undefined as any,
         strictProps: undefined as any,
         apiName: "Foo",
         description: "Foo interface desc",