+++ packages/generator/src/v2.0/generateClientSdkVersionTwoPointZero.test.ts
@@ -474,9 +474,13 @@ describe("generator", () => {
       export const $ontologyRid = 'ridHere';
       ",
         "/foo/index.ts": "export * from './ontology/actions';
+      export * as $Actions from './ontology/actions';
       export * from './ontology/interfaces';
+      export * as $Interfaces from './ontology/interfaces';
       export * from './ontology/objects';
+      export * as $Objects from './ontology/objects';
       export * from './ontology/queries';
+      export * as $Queries from './ontology/queries';
       export { $ontologyRid } from './OntologyMetadata';
       ",
         "/foo/ontology/actions.ts": "export { deleteTodos } from './actions/deleteTodos';
@@ -682,6 +686,7 @@ describe("generator", () => {
         ConvertProps as $ConvertProps,
         DefaultToFalse as $DefaultToFalse,
         FetchPageArgs as $FetchPageArgs,
+        IsAny as $IsAny,
         LinkedType as $LinkedType,
         LinkNames as $LinkNames,
         NullabilityAdherence as $NullabilityAdherence,
@@ -695,7 +700,7 @@ describe("generator", () => {
         ValidToFrom as $ValidToFrom,
       } from '@osdk/client.api';
 
-      export type OsdkObjectLinks$SomeInterface = never;
+      export type OsdkObjectLinks$SomeInterface = {};
 
       export namespace SomeInterface {
         export type PropertyKeys = 'SomeProperty';
@@ -727,7 +732,7 @@ describe("generator", () => {
             $PageResult<
               SomeInterface.OsdkObject<
                 (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
-                L
+                $IsAny<L> extends true ? SomeInterface.PropertyKeys : L
               >
             >
           >;
@@ -744,7 +749,7 @@ describe("generator", () => {
               $PageResult<
                 SomeInterface.OsdkObject<
                   (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
-                  L
+                  $IsAny<L> extends true ? SomeInterface.PropertyKeys : L
                 >
               >
             >
@@ -759,6 +764,7 @@ describe("generator", () => {
           osdkMetadata: typeof $osdkMetadata;
           objectSet: SomeInterface.ObjectSet;
           props: SomeInterface.Props;
+          linksType: OsdkObjectLinks$SomeInterface;
           strictProps: SomeInterface.StrictProps;
           description: 'Some interface';
           displayName: 'Sum Interface';
@@ -802,6 +808,7 @@ describe("generator", () => {
         osdkMetadata: $osdkMetadata,
         objectSet: undefined as any,
         props: undefined as any,
+        linksType: undefined as any,
         strictProps: undefined as any,
         apiName: 'SomeInterface',
         description: 'Some interface',
@@ -837,6 +844,7 @@ describe("generator", () => {
         ConvertProps as $ConvertProps,
         DefaultToFalse as $DefaultToFalse,
         FetchPageArgs as $FetchPageArgs,
+        IsAny as $IsAny,
         LinkedType as $LinkedType,
         LinkNames as $LinkNames,
         NullabilityAdherence as $NullabilityAdherence,
@@ -889,7 +897,7 @@ describe("generator", () => {
           ) => Promise<
             Person.OsdkObject<
               (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
-              L
+              $IsAny<L> extends true ? Person.PropertyKeys : L
             >
           >;
 
@@ -904,7 +912,7 @@ describe("generator", () => {
             $Result<
               Person.OsdkObject<
                 (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
-                L
+                $IsAny<L> extends true ? Person.PropertyKeys : L
               >
             >
           >;
@@ -920,7 +928,7 @@ describe("generator", () => {
             $PageResult<
               Person.OsdkObject<
                 (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
-                L
+                $IsAny<L> extends true ? Person.PropertyKeys : L
               >
             >
           >;
@@ -937,7 +945,7 @@ describe("generator", () => {
               $PageResult<
                 Person.OsdkObject<
                   (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
-                  L
+                  $IsAny<L> extends true ? Person.PropertyKeys : L
                 >
               >
             >
@@ -952,6 +960,7 @@ describe("generator", () => {
           osdkMetadata: typeof $osdkMetadata;
           objectSet: Person.ObjectSet;
           props: Person.Props;
+          linksType: Person.Links;
           strictProps: Person.StrictProps;
           description: 'A person';
           implements: [];
@@ -996,6 +1005,7 @@ describe("generator", () => {
         osdkMetadata: $osdkMetadata,
         objectSet: undefined as any,
         props: undefined as any,
+        linksType: undefined as any,
         strictProps: undefined as any,
         apiName: 'Person',
         description: 'A person',
@@ -1036,6 +1046,7 @@ describe("generator", () => {
         ConvertProps as $ConvertProps,
         DefaultToFalse as $DefaultToFalse,
         FetchPageArgs as $FetchPageArgs,
+        IsAny as $IsAny,
         LinkedType as $LinkedType,
         LinkNames as $LinkNames,
         NullabilityAdherence as $NullabilityAdherence,
@@ -1091,7 +1102,10 @@ describe("generator", () => {
             primaryKey: $PropertyValueClientToWire[Todo.Definition['primaryKeyType']],
             options?: $SelectArg<Todo.Definition, L, R, S>,
           ) => Promise<
-            Todo.OsdkObject<(S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'), L>
+            Todo.OsdkObject<
+              (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
+              $IsAny<L> extends true ? Todo.PropertyKeys : L
+            >
           >;
 
           readonly fetchOneWithErrors: <
@@ -1105,7 +1119,7 @@ describe("generator", () => {
             $Result<
               Todo.OsdkObject<
                 (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
-                L
+                $IsAny<L> extends true ? Todo.PropertyKeys : L
               >
             >
           >;
@@ -1121,7 +1135,7 @@ describe("generator", () => {
             $PageResult<
               Todo.OsdkObject<
                 (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
-                L
+                $IsAny<L> extends true ? Todo.PropertyKeys : L
               >
             >
           >;
@@ -1138,7 +1152,7 @@ describe("generator", () => {
               $PageResult<
                 Todo.OsdkObject<
                   (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
-                  L
+                  $IsAny<L> extends true ? Todo.PropertyKeys : L
                 >
               >
             >
@@ -1153,6 +1167,7 @@ describe("generator", () => {
           osdkMetadata: typeof $osdkMetadata;
           objectSet: Todo.ObjectSet;
           props: Todo.Props;
+          linksType: Todo.Links;
           strictProps: Todo.StrictProps;
           description: 'Its a todo item.';
           implements: ['SomeInterface'];
@@ -1214,6 +1229,7 @@ describe("generator", () => {
         osdkMetadata: $osdkMetadata,
         objectSet: undefined as any,
         props: undefined as any,
+        linksType: undefined as any,
         strictProps: undefined as any,
         apiName: 'Todo',
         description: 'Its a todo item.',
@@ -1443,9 +1459,13 @@ describe("generator", () => {
         export const $osdkMetadata = { extraUserAgent: '' };
         ",
           "/foo/index.ts": "export * from './ontology/actions.js';
+        export * as $Actions from './ontology/actions.js';
         export * from './ontology/interfaces.js';
+        export * as $Interfaces from './ontology/interfaces.js';
         export * from './ontology/objects.js';
+        export * as $Objects from './ontology/objects.js';
         export * from './ontology/queries.js';
+        export * as $Queries from './ontology/queries.js';
         ",
           "/foo/ontology/actions.ts": "export { deleteTodos } from './actions/deleteTodos.js';
         export { markTodoCompleted } from './actions/markTodoCompleted.js';
@@ -1650,6 +1670,7 @@ describe("generator", () => {
           ConvertProps as $ConvertProps,
           DefaultToFalse as $DefaultToFalse,
           FetchPageArgs as $FetchPageArgs,
+          IsAny as $IsAny,
           LinkedType as $LinkedType,
           LinkNames as $LinkNames,
           NullabilityAdherence as $NullabilityAdherence,
@@ -1663,7 +1684,7 @@ describe("generator", () => {
           ValidToFrom as $ValidToFrom,
         } from '@osdk/client.api';
 
-        export type OsdkObjectLinks$SomeInterface = never;
+        export type OsdkObjectLinks$SomeInterface = {};
 
         export namespace SomeInterface {
           export type PropertyKeys = 'SomeProperty';
@@ -1695,7 +1716,7 @@ describe("generator", () => {
               $PageResult<
                 SomeInterface.OsdkObject<
                   (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
-                  L
+                  $IsAny<L> extends true ? SomeInterface.PropertyKeys : L
                 >
               >
             >;
@@ -1712,7 +1733,7 @@ describe("generator", () => {
                 $PageResult<
                   SomeInterface.OsdkObject<
                     (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
-                    L
+                    $IsAny<L> extends true ? SomeInterface.PropertyKeys : L
                   >
                 >
               >
@@ -1727,6 +1748,7 @@ describe("generator", () => {
             osdkMetadata: typeof $osdkMetadata;
             objectSet: SomeInterface.ObjectSet;
             props: SomeInterface.Props;
+            linksType: OsdkObjectLinks$SomeInterface;
             strictProps: SomeInterface.StrictProps;
             description: 'Some interface';
             displayName: 'Sum Interface';
@@ -1770,6 +1792,7 @@ describe("generator", () => {
           osdkMetadata: $osdkMetadata,
           objectSet: undefined as any,
           props: undefined as any,
+          linksType: undefined as any,
           strictProps: undefined as any,
           apiName: 'foo.bar.SomeInterface',
           description: 'Some interface',
@@ -1805,6 +1828,7 @@ describe("generator", () => {
           ConvertProps as $ConvertProps,
           DefaultToFalse as $DefaultToFalse,
           FetchPageArgs as $FetchPageArgs,
+          IsAny as $IsAny,
           LinkedType as $LinkedType,
           LinkNames as $LinkNames,
           NullabilityAdherence as $NullabilityAdherence,
@@ -1857,7 +1881,7 @@ describe("generator", () => {
             ) => Promise<
               Person.OsdkObject<
                 (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
-                L
+                $IsAny<L> extends true ? Person.PropertyKeys : L
               >
             >;
 
@@ -1872,7 +1896,7 @@ describe("generator", () => {
               $Result<
                 Person.OsdkObject<
                   (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
-                  L
+                  $IsAny<L> extends true ? Person.PropertyKeys : L
                 >
               >
             >;
@@ -1888,7 +1912,7 @@ describe("generator", () => {
               $PageResult<
                 Person.OsdkObject<
                   (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
-                  L
+                  $IsAny<L> extends true ? Person.PropertyKeys : L
                 >
               >
             >;
@@ -1905,7 +1929,7 @@ describe("generator", () => {
                 $PageResult<
                   Person.OsdkObject<
                     (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
-                    L
+                    $IsAny<L> extends true ? Person.PropertyKeys : L
                   >
                 >
               >
@@ -1920,6 +1944,7 @@ describe("generator", () => {
             osdkMetadata: typeof $osdkMetadata;
             objectSet: Person.ObjectSet;
             props: Person.Props;
+            linksType: Person.Links;
             strictProps: Person.StrictProps;
             description: 'A person';
             implements: [];
@@ -1964,6 +1989,7 @@ describe("generator", () => {
           osdkMetadata: $osdkMetadata,
           objectSet: undefined as any,
           props: undefined as any,
+          linksType: undefined as any,
           strictProps: undefined as any,
           apiName: 'foo.bar.Person',
           description: 'A person',
@@ -2004,6 +2030,7 @@ describe("generator", () => {
           ConvertProps as $ConvertProps,
           DefaultToFalse as $DefaultToFalse,
           FetchPageArgs as $FetchPageArgs,
+          IsAny as $IsAny,
           LinkedType as $LinkedType,
           LinkNames as $LinkNames,
           NullabilityAdherence as $NullabilityAdherence,
@@ -2059,7 +2086,10 @@ describe("generator", () => {
               primaryKey: $PropertyValueClientToWire[Todo.Definition['primaryKeyType']],
               options?: $SelectArg<Todo.Definition, L, R, S>,
             ) => Promise<
-              Todo.OsdkObject<(S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'), L>
+              Todo.OsdkObject<
+                (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
+                $IsAny<L> extends true ? Todo.PropertyKeys : L
+              >
             >;
 
             readonly fetchOneWithErrors: <
@@ -2073,7 +2103,7 @@ describe("generator", () => {
               $Result<
                 Todo.OsdkObject<
                   (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
-                  L
+                  $IsAny<L> extends true ? Todo.PropertyKeys : L
                 >
               >
             >;
@@ -2089,7 +2119,7 @@ describe("generator", () => {
               $PageResult<
                 Todo.OsdkObject<
                   (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
-                  L
+                  $IsAny<L> extends true ? Todo.PropertyKeys : L
                 >
               >
             >;
@@ -2106,7 +2136,7 @@ describe("generator", () => {
                 $PageResult<
                   Todo.OsdkObject<
                     (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
-                    L
+                    $IsAny<L> extends true ? Todo.PropertyKeys : L
                   >
                 >
               >
@@ -2121,6 +2151,7 @@ describe("generator", () => {
             osdkMetadata: typeof $osdkMetadata;
             objectSet: Todo.ObjectSet;
             props: Todo.Props;
+            linksType: Todo.Links;
             strictProps: Todo.StrictProps;
             description: 'Its a todo item.';
             implements: ['foo.bar.SomeInterface'];
@@ -2182,6 +2213,7 @@ describe("generator", () => {
           osdkMetadata: $osdkMetadata,
           objectSet: undefined as any,
           props: undefined as any,
+          linksType: undefined as any,
           strictProps: undefined as any,
           apiName: 'foo.bar.Todo',
           description: 'Its a todo item.',
@@ -2555,6 +2587,7 @@ describe("generator", () => {
             ConvertProps as $ConvertProps,
             DefaultToFalse as $DefaultToFalse,
             FetchPageArgs as $FetchPageArgs,
+            IsAny as $IsAny,
             LinkedType as $LinkedType,
             LinkNames as $LinkNames,
             NullabilityAdherence as $NullabilityAdherence,
@@ -2576,7 +2609,7 @@ describe("generator", () => {
           export namespace UsesForeignSpt {
             export type PropertyKeys = 'id' | 'body';
 
-            export type Links = never;
+            export type Links = {};
 
             export interface Props {
               readonly body: $PropType['string'] | undefined;
@@ -2606,7 +2639,7 @@ describe("generator", () => {
               ) => Promise<
                 UsesForeignSpt.OsdkObject<
                   (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
-                  L
+                  $IsAny<L> extends true ? UsesForeignSpt.PropertyKeys : L
                 >
               >;
 
@@ -2621,7 +2654,7 @@ describe("generator", () => {
                 $Result<
                   UsesForeignSpt.OsdkObject<
                     (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
-                    L
+                    $IsAny<L> extends true ? UsesForeignSpt.PropertyKeys : L
                   >
                 >
               >;
@@ -2637,7 +2670,7 @@ describe("generator", () => {
                 $PageResult<
                   UsesForeignSpt.OsdkObject<
                     (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
-                    L
+                    $IsAny<L> extends true ? UsesForeignSpt.PropertyKeys : L
                   >
                 >
               >;
@@ -2654,7 +2687,7 @@ describe("generator", () => {
                   $PageResult<
                     UsesForeignSpt.OsdkObject<
                       (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
-                      L
+                      $IsAny<L> extends true ? UsesForeignSpt.PropertyKeys : L
                     >
                   >
                 >
@@ -2669,6 +2702,7 @@ describe("generator", () => {
               osdkMetadata: typeof $osdkMetadata;
               objectSet: UsesForeignSpt.ObjectSet;
               props: UsesForeignSpt.Props;
+              linksType: UsesForeignSpt.Links;
               strictProps: UsesForeignSpt.StrictProps;
               implements: [];
               interfaceMap: {};
@@ -2722,6 +2756,7 @@ describe("generator", () => {
             osdkMetadata: $osdkMetadata,
             objectSet: undefined as any,
             props: undefined as any,
+            linksType: undefined as any,
             strictProps: undefined as any,
             apiName: 'UsesForeignSpt',
             implements: [],
@@ -2891,9 +2926,13 @@ describe("generator", () => {
         export const $ontologyRid = 'ri.ontology.main.ontology.dep';
         ",
           "/foo/index.ts": "export * from './ontology/actions.js';
+        export * as $Actions from './ontology/actions.js';
         export * from './ontology/interfaces.js';
+        export * as $Interfaces from './ontology/interfaces.js';
         export * from './ontology/objects.js';
+        export * as $Objects from './ontology/objects.js';
         export * from './ontology/queries.js';
+        export * as $Queries from './ontology/queries.js';
         export { $ontologyRid } from './OntologyMetadata.js';
         ",
           "/foo/ontology/actions.ts": "export {};
@@ -2913,6 +2952,7 @@ describe("generator", () => {
           ConvertProps as $ConvertProps,
           DefaultToFalse as $DefaultToFalse,
           FetchPageArgs as $FetchPageArgs,
+          IsAny as $IsAny,
           LinkedType as $LinkedType,
           LinkNames as $LinkNames,
           NullabilityAdherence as $NullabilityAdherence,
@@ -2926,7 +2966,7 @@ describe("generator", () => {
           ValidToFrom as $ValidToFrom,
         } from '@osdk/client.api';
 
-        export type OsdkObjectLinks$SomeInterface = never;
+        export type OsdkObjectLinks$SomeInterface = {};
 
         export namespace SomeInterface {
           export type PropertyKeys = 'spt';
@@ -2958,7 +2998,7 @@ describe("generator", () => {
               $PageResult<
                 SomeInterface.OsdkObject<
                   (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
-                  L
+                  $IsAny<L> extends true ? SomeInterface.PropertyKeys : L
                 >
               >
             >;
@@ -2975,7 +3015,7 @@ describe("generator", () => {
                 $PageResult<
                   SomeInterface.OsdkObject<
                     (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
-                    L
+                    $IsAny<L> extends true ? SomeInterface.PropertyKeys : L
                   >
                 >
               >
@@ -2990,6 +3030,7 @@ describe("generator", () => {
             osdkMetadata: typeof $osdkMetadata;
             objectSet: SomeInterface.ObjectSet;
             props: SomeInterface.Props;
+            linksType: OsdkObjectLinks$SomeInterface;
             strictProps: SomeInterface.StrictProps;
             displayName: 'Sum Interface';
             implements: [];
@@ -3031,6 +3072,7 @@ describe("generator", () => {
           osdkMetadata: $osdkMetadata,
           objectSet: undefined as any,
           props: undefined as any,
+          linksType: undefined as any,
           strictProps: undefined as any,
           apiName: 'com.example.dep.SomeInterface',
           displayName: 'Sum Interface',
@@ -3062,6 +3104,7 @@ describe("generator", () => {
           ConvertProps as $ConvertProps,
           DefaultToFalse as $DefaultToFalse,
           FetchPageArgs as $FetchPageArgs,
+          IsAny as $IsAny,
           LinkedType as $LinkedType,
           LinkNames as $LinkNames,
           NullabilityAdherence as $NullabilityAdherence,
@@ -3083,7 +3126,7 @@ describe("generator", () => {
         export namespace Task {
           export type PropertyKeys = 'taskId' | 'body';
 
-          export type Links = never;
+          export type Links = {};
 
           export interface Props {
             readonly body: $PropType['string'] | undefined;
@@ -3111,7 +3154,10 @@ describe("generator", () => {
               primaryKey: $PropertyValueClientToWire[Task.Definition['primaryKeyType']],
               options?: $SelectArg<Task.Definition, L, R, S>,
             ) => Promise<
-              Task.OsdkObject<(S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'), L>
+              Task.OsdkObject<
+                (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
+                $IsAny<L> extends true ? Task.PropertyKeys : L
+              >
             >;
 
             readonly fetchOneWithErrors: <
@@ -3125,7 +3171,7 @@ describe("generator", () => {
               $Result<
                 Task.OsdkObject<
                   (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
-                  L
+                  $IsAny<L> extends true ? Task.PropertyKeys : L
                 >
               >
             >;
@@ -3141,7 +3187,7 @@ describe("generator", () => {
               $PageResult<
                 Task.OsdkObject<
                   (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
-                  L
+                  $IsAny<L> extends true ? Task.PropertyKeys : L
                 >
               >
             >;
@@ -3158,7 +3204,7 @@ describe("generator", () => {
                 $PageResult<
                   Task.OsdkObject<
                     (S extends false ? '$notStrict' : never) | ($DefaultToFalse<R> extends false ? never : '$rid'),
-                    L
+                    $IsAny<L> extends true ? Task.PropertyKeys : L
                   >
                 >
               >
@@ -3173,6 +3219,7 @@ describe("generator", () => {
             osdkMetadata: typeof $osdkMetadata;
             objectSet: Task.ObjectSet;
             props: Task.Props;
+            linksType: Task.Links;
             strictProps: Task.StrictProps;
             implements: [];
             interfaceMap: {};
@@ -3218,6 +3265,7 @@ describe("generator", () => {
           osdkMetadata: $osdkMetadata,
           objectSet: undefined as any,
           props: undefined as any,
+          linksType: undefined as any,
           strictProps: undefined as any,
           apiName: 'com.example.dep.Task',
           implements: [],