+++ packages/api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @osdk/api
 
+## 1.10.0-beta.1
+
+### Minor Changes
+
+- dc25fb4: Fixes edge cases in inferred return types especially when changing object types
+
 ## 1.10.0-beta.0
 
 ### Minor Changes