+++ packages/client.test.ontology/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @osdk/client.test.ontology
 
+## 1.2.0-beta.5
+
+### Patch Changes
+
+- Updated dependencies [dc25fb4]
+  - @osdk/client.api@0.22.0-beta.5
+  - @osdk/api@1.10.0-beta.1
+
+## 1.2.0-beta.4
+
+### Patch Changes
+
+- Updated dependencies [0ea1a17]
+  - @osdk/client.api@0.22.0-beta.4
+
 ## 1.2.0-beta.3
 
 ### Patch Changes