+++ packages/e2e.sandbox.todoapp/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @osdk/e2e.sandbox.todoappapp
 
+## 3.0.0-beta.5
+
+### Patch Changes
+
+- Updated dependencies [dc25fb4]
+  - @osdk/client.api@0.22.0-beta.5
+  - @osdk/client@0.22.0-beta.5
+  - @osdk/api@1.10.0-beta.1
+
+## 3.0.0-beta.4
+
+### Patch Changes
+
+- Updated dependencies [0ea1a17]
+  - @osdk/client.api@0.22.0-beta.4
+  - @osdk/client@0.22.0-beta.4
+
 ## 3.0.0-beta.3
 
 ### Patch Changes